Runc and version artifacts
Former-commit-id: f443f0b326cd94f52e8da8482b952088318a221e

+++ podman-update/00-podman-update.sh
@@ -2,23 +2,41 @@
 
 mkdir -p /mnt/data/.cache
 
-PODMAN_DL=/mnt/data/.cache/podman
-CONMON_DL=/mnt/data/.cache/conmon
+PODMAN_VERSION=3.2.0-dev
+RUNC_VERSION=1.0.0-rc95
+CONMON_VERSION=2.0.27
+PODMAN_DL=/mnt/data/.cache/podman-$PODMAN_VERSION
+RUNC_DL=/mnt/data/.cache/runc-$RUNC_VERSION
+CONMON_DL=/mnt/data/.cache/conmon-$CONMON_VERSION
 SECCOMP=/usr/share/container/seccomp.json
 
 if [ ! -f CONMON_DL ]; then
-  curl -fsSLo $CONMON_DL https://github.com/containers/conmon/releases/download/v2.0.27/conmon.arm64
+  curl -fsSLo $CONMON_DL https://github.com/containers/conmon/releases/download/v$CONMON_VERSION/conmon.arm64
 fi
 chmod +x $CONMON_DL
-mv /usr/libexec/podman/conmon /usr/libexec/podman/conmon.old
+if [ ! -f /usr/libexec/podman/conmon.old ]; then
+  mv /usr/libexec/podman/conmon /usr/libexec/podman/conmon.old
+fi
+
 ln -s $CONMON_DL /usr/libexec/podman/conmon
 
 if [ ! -f $PODMAN_DL ]; then
-  curl -fsSLo $PODMAN_DL https://raw.githubusercontent.com/boostchicken/udm-utilities/master/podman-update/bin/podman-3.2.0-dev
+  curl -fsSLo $PODMAN_DL https://raw.githubusercontent.com/boostchicken/udm-utilities/master/podman-update/bin/podman-$PODMAN_VERSION
 fi
 chmod +x $PODMAN_DL
-mv /usr/bin/podman /usr/bin/podman.old
+if [ ! -f /usr/bin/podman.old ]; then
+  mv /usr/bin/podman /usr/bin/podman.old
+fi
 ln -s $PODMAN_DL /usr/bin/podman
+
+if [ ! -f $RUNC_DL ]; then
+  curl -fsSLo $RUNC_DL https://raw.githubusercontent.com/boostchicken/udm-utilities/master/podman-update/bin/runc-$RUNC_VERSION
+fi
+chmod +x $RUNC_DL
+if [ ! -f /usr/bin/runc.old ]; then
+  mv /usr/bin/runc /usr/bin/runc.old
+fi
+ln -s $RUNC_DL /usr/bin/runc
 
 if [ ! -f $SECCOMP ]; then
   mkdir -p /usr/share/container/
